Basic UI completed
