+++ src/components/todo/ErrorComponent.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function ErrorComponent() {
+    return <div>Error</div>
+}
+
+export default ErrorComponent;
