+++ src/components/todo/FooterComponent.jsx
@@ -1,0 +1,13 @@
+import React, {Component} from 'react';
+
+class FooterComponent extends Component {
+    render() {
+        return (
+            <footer className="footer">
+                <span className="text-muted">All Rights Reserved 2024 @Andrew</span>
+            </footer>
+        )
+    }
+}
+
+export default FooterComponent;
