+++ src/components/todo/TodoApp.jsx
@@ -4,20 +4,30 @@ import withNavigation from './WithNavigation.jsx'
 import withParams from './withParams.jsx'
 import '../../App.css';
 import '../../bootstrap.css';
+import AuthenticatedRoute from './AuthenticatedRoute.jsx';
+import LoginComponent from './LoginComponent.jsx';
+import ListTodosComponent from './ListTodosComponent.jsx';
+import ErrorComponent from './ErrorComponent.jsx';
+import HeaderComponent from './HeaderComponent.jsx';
+import FooterComponent from './FooterComponent.jsx';
+import LogoutComponent from './LogoutComponent.jsx';
+import WelcomeComponent from './WelcomeComponent.jsx';
 
 class TodoApp extends Component {
     render() {
         const LoginComponentWithNavigation = withNavigation(LoginComponent);
         const WelcomeComponentWithParams = withParams(WelcomeComponent);
+        const HeaderComponentWithNavigation = withNavigation(HeaderComponent);
+
         return (
             <div className="TodoApp">
                 <Router>
-                    <HeaderComponent />
+                    <HeaderComponentWithNavigation/>
                     <Routes>
                         <Route path="/" element={<LoginComponentWithNavigation />} />
                         <Route path="/login" element={<LoginComponentWithNavigation />} />
-                        <Route path="/welcome/:name" element={<WelcomeComponentWithParams />} />
-                        <Route path="/todos" element={<ListTodosComponent />} />
+                        <Route path="/welcome/:name" element={<AuthenticatedRoute><WelcomeComponentWithParams /></AuthenticatedRoute>} />
+                        <Route path="/todos" element={<AuthenticatedRoute><ListTodosComponent /></AuthenticatedRoute>} />
                         <Route path="/logout" element={<LogoutComponent />} />
                         <Route path="*" element={<ErrorComponent />} />
                     </Routes>
@@ -27,183 +37,5 @@ class TodoApp extends Component {
         )
     }
 }
-
-class FooterComponent extends Component {
-    render() {
-        return (
-            <footer className="footer">
-                <span className="text-muted">All Rights Reserved 2024 @Andrew</span>
-            </footer>
-        )
-    }
-}
-
-class LogoutComponent extends Component {
-    render() {
-        return (
-            <div>
-                <h1>You are logged out</h1>
-                <div className="container">
-                    Thank you for using our application.
-                </div>
-            </div>
-        )
-    }
-}
-
-class HeaderComponent extends Component {
-    render() {
-        return (
-            <header>
-                <nav className="navbar navbar-expand-md navbar-dark bg-dark">
-                    <ul className="navbar-nav">
-                        <li><Link className="nav-link" to="/welcome/Andrew">Home</Link></li>
-                        <li><Link className="nav-link" to="/todos">Todos</Link></li>
-                    </ul>
-                    <ul className="navbar-nav navbar-collapse justify-content-end">
-                        <li><Link className="nav-link" to="/login">Login</Link></li>
-                        <li><Link className="nav-link" to="/logout">Logout</Link></li>
-                    </ul>
-                </nav>
-            </header>
-        )
-    }
-}
-
-class ListTodosComponent extends Component {
-    constructor(props) {
-        super(props)
-        this.state = {
-            todos: [
-                {id: 1, description: 'Learn AWS', done: false, targetDate: new Date()},
-                {id: 2, description: 'Learn DevOps', done: false, targetDate: new Date()},
-                {id: 3, description: 'Learn Full Stack Development', done: false, targetDate: new Date()}
-            ]
-        }
-    }
-    render() {
-        return (
-            <div>
-                <h1>List of Todos</h1>
-                <div className="container">
-                    <table className="table">
-                        <thead>
-                            <tr>
-                                <th>Description</th>
-                                <th>Is Done?</th>
-                                <th>Target Date</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            {this.state.todos.map(todo => (
-                                <tr key={todo.id}>
-                                    <td>{todo.description}</td>
-                                    <td>{todo.done.toString()}</td>
-                                    <td>{todo.targetDate.toString()}</td>
-                                </tr>
-                            ))}
-                        </tbody>
-                    </table>
-                </div>
-            </div>
-        )
-    }
-}
-
-class WelcomeComponent extends Component {
-    render() {
-        return (
-            <div>
-                <h1>Welcome!</h1>
-                <div className="container">
-                    Welcome {this.props.params.name}. You can manage your todos <Link to="/todos">here</Link>.
-                </div>
-            </div>
-        )
-    }
-}
-
-function ErrorComponent() {
-    return <div>Error</div>
-}
-
-class LoginComponent extends Component {
-
-    constructor (props) {
-        super(props)
-        this.state = {
-            username: 'Andrew O',
-            password: '',
-            hasLoginFailed: false,
-            showSuccessMessage: false
-        }
-
-        // this.handleUsernameChange = this.handleUsernameChange.bind(this)
-        // this.handlePasswordChange = this.handlePasswordChange.bind(this)
-        this.handleChange = this.handleChange.bind(this)
-        this.loginClicked = this.loginClicked.bind(this)
-    }
-
-    handleChange(event) {
-        //console.log(this.state)
-        this.setState({[event.target.name]:event.target.value})
-    }
-
-    loginClicked() {
-        //Username: Andrew O
-        //Password: test
-        if(this.state.username === 'Andrew O' && this.state.password === 'test'){
-            this.props.navigate(`/welcome/${this.state.username}`)
-            //this.setState({showSuccessMessage:true})
-            //this.setState({hasLoginFailed:false})
-        }
-        else{
-            console.log('Failed')
-            this.setState({showSuccessMessage:false})
-            this.setState({hasLoginFailed:true})
-        }
-    }
-
-    // handleUsernameChange(event) {
-    //     console.log(event.target.value)
-    //     this.setState({username:event.target.value})
-    // }
-
-    // handlePasswordChange(event) {
-    //     console.log(event.target.value)
-    //     this.setState({password:event.target.value})
-    // }
-
-    render () {
-        return (
-            <div>
-                <h1>Login</h1>
-                <div className="container">
-                    {/*<ShowInvalidCredentials hasLoginFailed={this.state.hasLoginFailed}/>*/}
-                    {this.state.hasLoginFailed && <div className="alert alert-warning">Invalid Credentials</div>}
-                    {this.state.showSuccessMessage && <div>Login Successful</div>}
-                    {/*<ShowLoginSuccessMessage showSuccessMessage={this.state.showSuccessMessage}/>*/}
-                    User Name: <input type="text" name="username" value={this.state.username} onChange={this.handleChange}/>
-                    Password: <input type="password" name="password" value={this.state.password} onChange={this.handleChange}/>
-                    <button className="btn btn-success" onClick={this.loginClicked}>Login</button>
-                </div>
-            </div>
-        )
-    }
-}
-
-// function ShowInvalidCredentials(props){
-//     if(props.hasLoginFailed){
-//         return <div>Invalid Credentials</div>
-//     }
-//     return null
-// }
-
-// function ShowLoginSuccessMessage(props){
-//     if(props.showSuccessMessage){
-//         return <div>Login Successful</div>
-//     }
-//     return null
-// }
 
 export default TodoApp
