+++ src/components/todo/WelcomeComponent.jsx
@@ -1,0 +1,17 @@
+import React, {Component} from 'react';
+import {Link} from 'react-router-dom';
+
+class WelcomeComponent extends Component {
+    render() {
+        return (
+            <div>
+                <h1>Welcome!</h1>
+                <div className="container">
+                    Welcome {this.props.params.name}. You can manage your todos <Link to="/todos">here</Link>.
+                </div>
+            </div>
+        )
+    }
+}
+
+export default WelcomeComponent;
